Revert "Fix DeviceConfig resetting."
This change went into internal after R branched, and did not
go to AOSP. Revert it in mainline-prod in order to ensure that
CtsHostsideNetworkTests builds.

These tests are developed in internal, not AOSP, so do not need
to automerge to mainline-prod.

Bug: 167645754
Test: treehugger
Ignore-AOSP-First: needed to sync mainline-prod with AOSP
Change-Id: Ieb1f0ce2a69180bbe89f810ddb3b48188ae3530e
Merged-In: Ib0ac49609e444a53a6fee4575f5078e15f364eef