CTS to ensure telecom can bind to NonUiInCallService wo export prop.
This CL is a postmortem action which is detailed in the linked bug.
In short, an OEM had the telecom stack outside of the system process
and had a non-ui InCallService it wanted to bind.  However, the
InCallService set the Exported property to false.  This combo caused
a calling outage that prevented users from making/receiving calls.

To prevent the above scenaro from occuring again, this new CTS test
was created. It tests the Telecom stack of the OEM against a
non-ui InCallService that has it's exported property set to false.

Bug: 198715680

Test: 1 new CTS test,
      NonUiInCallServiceWoExportTest
      #testTelecomCanBindToNonUiInCallServiceThatIsNotExported

Change-Id: I233de7ba23a7940b502362f1557e97a26e96b7c2