Do not keep WearableExtender background bitmaps
setBackground() call was deprecated in API level 29 - however we
still keep the background bitmap in memory - both in system server
and systemui - if the app decides to set it. This is a rather big
contributor to Notification memory use.

This change makes setBackground() call a no-op on devices running API 35
or newer.

Bug: 270551184
Test: newly created CTS and unit tests run on cheetah
Change-Id: Ib11e07ee0d887a361fc4ce018b3034751559117f