Refactored testCropWithChildBoundingInsets to avoid recording
testCropWithChildBoundingInsets only needs a single frame since there is
no animation running. Move testCropWithChildBoundingInsets to
AttachedSurfaceControlTest and capture a screenshot of the activity when
everything is set up. This avoids the overhead of setting up a
screenrecording and avoids any issues system content on screen that
makes the test flaky.

Test: AttachedSurfaceControlTest#testCropWithChildBoundingInset
Bug: 282169297
Change-Id: I0211d50d6cb5b6a44655084da3dfd4b7f97b5880