Test that TextClassifier PI can start activity
This tests that the default listener of a TextClassification can start a
background activity. Unfortunately it is impossible to test this for
listeners implemented/provided by the app developer as the code that
sends the PI (and needs to allow background activity launches) is in
the listener itself.

This is a followup to a p0 fix (b/284621682) and it only adds test
logic. No risk of adding instability, but might prevent regressions in
future fixes/backports.

Bug: 286004957
Change-Id: Ic5aea95613b55ae89e9f61eb7e6babefbf9745d2
Test: atest BackgroundActivityLaunchTest