Optimize testLongProcessingRepeatingFlexibleYuv
The test currently runs for all available sizes. There are 30 available
sizes that the test runs on and each size has 10 frames to verify the
long processing time (which takes ~3-4 seconds). That is ~90-120s in
practice per camera. The test is run on 2 cameras - 0 and 1, which makes
the total running time ~200s.

Since the sizes for the given test are not that important, it is
sufficient to run it on the min and max sizes from available sizes,
which reduces the test time significantly, while maintaining the coverage.

The test time for Oriole on the latest stable build (10337921) goes
down from ~3m20s to ~14s (~93% improvement).

Bug: 287260354
Test: atest ImageReaderTest#testLongProcessingRepeatingFlexibleYuv
Change-Id: Iea320c692f090f94bb36bee53919eb88dac66f13