TouchModeTest: Improve robustness of waiting for window losing focus
When we expect a window to gain focus using WindowUtil#waitForFocus, it
waits for a generous 10 secs, while when we wait for the window to lose
focus, we use the PollingCheck default of 3 seconds.

Use the same polling timeout for window losing focus. Also, wait for app
transitions to idle beforehand to improve robustness.

Bug: 287443739
Test: atest TouchModeTest
Change-Id: Ia9a12db6503f97b36519de5cfccf97e8cdcf1a03