Setup airplane mode to disable GPP for CtsSilentUpdateHostTestCases
In this test suite, we would like test apps (SilentInstallR and
SilentInstallS) that explicitly target old SDKs. Instead of enabling
verifier_engprod to reduce GPP timeout, the test needs to fully
disable GPP. To diable GPP, we will turn on airplane mode before the
test and turn off the airplane mode after the test to keep the device
in the good state for next tests.

Bug: 288337948
Test: atest SilentUpdateHostsideTests#silentInstallRepeatedly_RequiresUserAction
Test: atest SilentUpdateHostsideTests#newInstallDpc_withInstallDpcPermission_requiresNoUserAction
Test: atest SilentUpdateHostsideTests#updateSApp_RequiresNoUserAction
Change-Id: I771da711edaec2b8838d3ea343a91d7013cf141e