Choreographer CTS fix for higher refresh rate.
The test assertion is ensuring that api users will always have a
baseline number of frame timeline choices, by enforcing choices to span
a certain duration.
Currently, the code in the platform uses 7 choices. However, for higher
frame rate devices such as 144Hz, the deadline is not 50 ms in the
future but 1/144*7 ~ 48.6ms. Therefore change the duration to 45 ms in
the test for now.

Bug: 291121167
Test: atest ChoreographerNativeTest
Change-Id: I23135f96ef7b06eaf7cae8b325dd907dd79b6509