Remove testAppOrientationRequestConfigClears
It is expected to keep the last device orientation if the sensor
cannot decide orientation, i.e. return -1 as flat rotation or there
has not been a sensor event reported.

When sensor rotation is -1:
1. Launch an unspecified orientation activity from home.
   Home is portrait, so it is also portrait.
2. Launch landscape activity.
3. Launch the unspecified orientation activity.
   Previous activity is landscape, so it is also landscape.
But the tests asserts that the last activity should be portrait.
The condition is only satisfied if sensor reported a specific
rotation. So there is no strong reason to enforce it.

Also remove the config-change assertion of
testTaskCloseRestoreFreeOrientation because the test activity may
receive config change while it was launched in portrait but rotates
to landscape when returning from a landscape activity.

Bug: 292086599
Test: AppConfigurationTests#testTaskCloseRestoreFreeOrientation
Change-Id: Ice2d9d8dd118cbc2486aae4078db37b7592a15e3