Fix flaky database test
Bug: 292296448

AbstractCursorTest#testSetNotificationsUris() is failing about 2% of
the time.  The root cause is that the test app is sometimes put in the
CACHED_EMPTY state, and this causes all notifications to be delayed by
10s.  The 10s delay exceeds the 5s timeout hard-coded in the test.

As a temporary fix, the timeout is extended from 5s to 15s.  A
separate bug has been filed to look for a solution that removes the
occasional 10s delay completely.

The test was run 200 times to verify that the flakiness is gone.

Test: atest
 * CtsDatabaseTestCases

Change-Id: I353ee07ddaeb1235dc9d03ef0bb5638433650e60