Only run flaky permissions test in postsubmit (CTS)
The folder cts/tests/tests/permissionui contains a TEST_MAPPING
which runs CtsPermissionUiTestCases in presubmit. Move it to
postsubmit.

The reason for doing this is part of a larger effort:

Certain tests in these test modules have a flake rate higher than 0.5%,
which means they are out-of-SLO for Project Ironwood. We've marked them
as @FlakyTest (in a separate change). But, in this change, we also
update the TEST_MAPPING files so that don't run any flaky tests
(i.e., annotated with @FlakyTest) in presubmit, but do run them in
postsubmit.

To test: For each affected TEST_MAPPING, run "atest :X --dry-run",
where X is each affected test-group, e.g., ("presubmit-large",
"presubmit", "postsubmit","mainline-presubmit","mainline-postsubmit")

Bug: 290240321
Test: see above
Change-Id: I826c6758eb5f4af32540453bfeb8fca0ab411fa3