Implement the BD Rate Calculator
Tests are provided on this implementation that use the same
data as the libvmaf implementation and ensure this implementation
is consistent.

There is a slight difference in the result caused by the difference
in interpolation methods, but in-general the results hold and should
provide an accurate calculation of BD rate.

Bug: 290290959
Test: Unit Tests provided tied to known-good data.
Change-Id: Iee91d106f95ad2fc1e9915c318938aeeb60aa451