InputAtomsTest: split counts into separate tests
We were getting a number of flakes where pinch or three-finger swipe
gestures weren't getting counted correctly. Splitting the test into four
tests (one for each gesture type) and adding a delay after each gesture
is performed but before the device is closed (to allow time for the end
of the gesture to be reported and counted) reduces the flakiness
significantly.

Bug: 293373752
Test: run the tests for 500 iterations or so without failure
Change-Id: I3b8fa0e183a89f2e296f9fafb1437b0c933f6e8d