Defalke testContextCommitted_noSaveUi_whileEmptyValueForRequiredIds  by explicitly focus on password field
Originally the test is asserting there is only one data_type_shown event. It counts on the changing text on password field is fast so it assumes there is not a second dropdown shown. So the original test has potential for flakiness.

The fix is to have the Ui Bot focus on password field and wait for idle sync before setting the text. In this way, the dataset will have time to
show.

The four fill events the test is asserting are DATATYPE_SHOWN_EVENT, VIEW_REQUEST_AUTOFILL, DATATYPE_SHOWN_EVENT, CONTEXT_COMMIT_EVENT.

Bug:b/294461217
Test:atest CtsAutoFillServiceTestCases

Change-Id: Ic6e2bb6215d8fb2496c638b4d64de76e8c805733