[Autofill CTS] Add Cts test for saveUi flash fix
Add a new test that verifies the fix for saveUi flashing. The test will
trigger the saveUi when it presses the login button. At the same time,
a new view will appear in the same activity, triggering autofill
request. The test will verify that the subsequent autofill response will
not close the saveUi.

Test: atest android.autofillservice.cts.dropdown.LoginActivityTest#testSaveUiWhenAutoFocusedToANewEditText
Bug: 283282918
Change-Id: I56bdd16123dbcfaa7e0aa93804c5b1a746a6398c