Revert "Add tests for aborbing scroll flings"
This reverts commit b381f5ac64e2b7d8fee28c2a4159118759a894e3.

Reason for revert: The tests are poorly written.

Fixes: 294958365

The two tests added are not really CTS tests, but functional tests.
They also improperly check that onRelease() is called when the
onPullDistance() and onRelease() shouldn't be called at all after
a fling. The tests are low quality and should be removed.

Test: ran ScrollViewTest

Change-Id: I58fd3ac45029936d0732e98fdca39fa15f5ab8f6