Reduce timeouts in VDM CTS
Previously the tests attempted to launch an Activity and checked back
after a N second sleep that the Activity was not launched.

This logic is replaced with logic that waits for the
BlockedAppStreamingActivity to be launched by GWPC. And once it's
launched it verifies that the Activity the test attempted to launch was
not started either.

This reduces the amount of time spent waiting by multiple seconds for
each affected test (N second timeout down to hundreds of ms).

Bug: 292050907
Test: atest CtsVirtualDevicesAppLaunchTestCases
CtsVirtualDevicesSensorTestCases

Change-Id: Icf7d853ec5b060dea9a0f61e98b24b075f6d8787