Introduce WakeUpAndUnlockRule
Adds a TestRule that ensures device is awake and unlocked before running
the test.
Also moves the wakeUpAndUnlock function into UiDeviceUtils to not have
the test rule call a static method in the base test class which feels a
bit weird.

Why TestRule? Because wakeUpAndUnlock uses Shell permissions and can
interfere with other usages of Shell permissions, especially when using
AdoptShellPermissionsRule. With this, both can be wrapped nicely into a
RuleChain.

Bug: 294967537
Test: atest + child change
Change-Id: I2262d702f2f4f18db251d1ae7bae32edbb6cfab0