Log windowInsets received and mark test Flaky
This is a manual cherry-pick of the CL from main, as the file location
changed between branches.

This CL marks the testDispatchApplyWindowInsetsCount_ime as Flaky, and
adds logging for the received windowInsets to aid in finding the root
cause.

Test: n/a
Bug: 297000797
Merged-In: I2e30af2d9a131aeaf3a82875318c1ca995517cbf
Change-Id: I12de9c7d79a7077b1441011c8a038a3e1ebf9270