Don't assume incremental fs exists on R launch devices
With R launch devices, OEMs were allowed to either ship incfs as a
module or built in to the kernel. If they chose to build it in, then
when we test GSI/GKI, incfs is not present so incfs tests all fail.

Since we cannot test for GSI and R launch combo, do not assume incfs on
R launch devices

Bug: 287545585
Test: atest CtsIncrementalInstallHostTestCases passes on a S or above
device

Change-Id: I1c6c7bf15eee5a6f1a1eb6fd3381816d73550413