Camera CTS: Address gap in CTS for external camera testing
- Currently, CTS only includes a single check to validate the connection
status of an external camera. However, external camera CTS testing is
easy to bypass by disconnecting the external camera for the remaining
CTS test run after the initial validation step.
- Update the retrieval of camera IDs under test to using a method that
verifies the external camera connection if the external camera feature
is supported, as it should be included among the cameras to be tested.

Test: Camera CTS with and without external feature flag enabled
Bug: 238240477
Change-Id: If687d70567437e3623b9879b6c354a85f1782b4e