Update color listener CTS for system only
Before, in case of a migration (initial state: shared wallpaper on
system + lock, then: set a wallpaper on system only), we would notify
that the colors changed on FLAG_LOCK. Actually it is not the case, in
case of migration the colors on FLAG_LOCK do not change. It is a
performance improvement to not notifyWallpaperColorsChanged in this
case.

This CL updates the CTS to match this behaviour. The test does not
change if the LWP flag is not enabled.

Flag: lockscreen LWP
Bug: 298366419
Test: atest WallpaperManagerTest

Change-Id: I50869a86d3200cff1cf88703fad1184d319b954e