Run shell command as root to access secondary user files.
On HSUM (Headless System User Mode) devices, bugreport is generated for
a secondary user, and `ls` shell command was failing to access the
bugreport file.
This change runs the shell command as root when the file is not owned by
the system user (user 0). Running a shell command as root only works on
a debug build. Therefore we are skipping some tests on `-user` build for
now.

Bug: 296720745

Test: atest android.bugreport.cts.BugreportManagerTest

Change-Id: I3387709803ba52fa187178e4d6a20ac966b5b795