Skip assist no data test for auto
Currently automotive test requires data and content to be present as
multi-displays and secure displays prevent data from being
delievered to the voice interaction session. The data is used to
identify if the information at the main voice interaction session
is correct, thus having no data essentially voids these checks.
Filed a separate bug (b/299988169) to follow up on providing some
mechanism for display info in the framework to properly fix the test
for multi-displays and secure displays.

Bug: 288763616
Test: atest ExtraAssistDataTest
Change-Id: I7aa0d00eaf8a184697a196c869e259d8e552fc8a