CtsContentTestCases: replace more configVarying type with strings and styles
The configVarying type is only used in this test, and causes problems
for downstream tools that expect to know all the valid type.  Replace
the custom type with strings and styles.

This is the same as Ib1895f7c7d37a9796cd6a8adb6ae416ad55205e2 for some
newly introduced uses of R.configVarying.

Bug: 294256649
Test: atest CtsContentTestCases
Change-Id: Ib1895f7c7d37a9796cd6a8adb6ae416ad55205e2