Add a delay before moving on with the test.
The test passes locally on tablets, but it doesn't change orientation.
Instead, black bars are rendered. Nonetheless the test passes.

The test passes on manual runs in ATP on tablets too.

The test fails on cuttlefish with automated tests.
This CL tries to see if it's simply a race condition issue of the
state of the system not propagating on time. Adding extra wait after
entering PiP. The logs show there is mismatch in display bounds indicating that it is in the wrong orientation as the test fails.

To be reverted once more information is found from debugging.

Bug: 300046278
Change-Id: I633458b3da82659189ce6dbc7867cbbdeb4a0cd3
Test: atest PinnedStackTests