Skip wallpaper visibility check when lockscreen_live_wallpaper is off
When the flag lockscreen_live_wallpaper is not enabled, the visibility
of the wallpaper on the lockscreen behind a translucent activity is not well defined in the case of two separate wallpapers. We are then going to remove this check to let the test pass when separate wallpapers are set.

Fix: 296190127
Test: KeyguardTests
Change-Id: I32414e07b72401cedfda9fbca4a33fc645fcd73e