Remove the SQLite version test from main
There are multiple versions of SQLite in the current Android branches,
but all are being tested with a single branch.  Therefore, the SQLite
version test can never work: if it works for one branch it will fail
for another.  This CL removes the version check entirely.  The version
check will be restored when the test-branch/image-branch pairing is
better understood.

Test: atest
 * CtsDatabaseTestCases

Bug: 308213196
Change-Id: I460ef77cecb46546cf62a80b12ee249f57f57d0d