Reduce redundant waiting
Waiting for the activity to RESUMED state before checking the
activity valid state. With that, we can skip waiting for surface
shown because the #waitForValidState also waits the activity
window to be visible.

The original test flakiness could be due to the test starts
an activity that has the same component name as the current top
activity and therefore pass through #waitForValidState since
the original top activity has all the valid states.

Bug: 293749685
Test: atest IntentTests
Change-Id: Icf03aebc491aa14b0ac9425aa31e167e016e045a