Modifying ActionGetContentOnlyTest#findAndClickMediaIcon to work with multiple app_labels.
While solving an accessibility bug the app_label was modified from 'Media' to 'Media picker' and after making the modification, since this test had the hardcoded value for the name as 'Media' it started failing. After fixing this some versions of the code became incompatible with this test and hence have modified the test to work with both names.

Test: atest ActionGetContentOnlyTest.java

Bug: 301214478
Change-Id: I83d1ef31d265fdb93e5e8e1f895392500a30c0bf