PersistableBundle for Ravenwood, with CTS.
We have the XML interfaces from libcore, but the implementations
there are tied closely to things like StringPool.  We thankfully
have our own human-readable XML serializer that we could use, but
there's unfortunately not a good pull parser.

To get us unstuck for the moment, pivot the internals to
unconditionally use the binary XML format, which at least we
know is thoroughly tested, and any data written through this path
can then also be parsed.

This is enough to support PersistableBundle, and we'll circle back
to get ourselves a human-readable XML implementation in the future.

Bug: 292141694
Test: atest-dev CtsOsTestCasesRavenwood CtsOsTestCases
Change-Id: I0810d962f66afbba84441ee075119dc81af85542