Fix private profile disable quiet mode test
The test on disabling quiet mode for private profile currently expects
to receive the profile available broadcast. However, as we moved to
using the property isAuthAlwaysRequriedToDisableQuietMode, we would need
to check for credentials as a requirement to disable quiet mode. The
test now fails since it expects the broadcast to be received. With this
change, we would move to checking that the requestQuietModeEnabled API
returns false when we try to disable quiet mode for private profiles.
The API is meant to return false when the user's credentials are
required to be checked to disable quiet mode, which is also whats
required to check.

Test: atest UserManagerTest
Bug: 311179780
Change-Id: I12f9ad777268aa2525ae72dd3cd55fe8f1ac04a8