[DO NOT MERGE] Fix testWindowProducerScaling CTS issue
Swap the workaround (I9f5c97f75208604796eb12e1086c56f30b91f8ab) with a
proper solution.

Disable scrims, which are shown on top of the status and navigation bar
when the bars are transparent. The scrims affect PixelCopyTest suite
when running on builds which use "classic" navigation bar (unlike, for
example, phone builds, where gestures are used and the nav bar is
effectively non-existent).

Note: the fix was previously submitted to one of the newer branches, see
more details here - b/300268614#comment23. This change is change is a
port back of that fix, where the previously submitted workaround is
removed.

Bug: 300268614
Fix: 310801230
Test: atest-src android.view.cts.PixelCopyTest -c
Change-Id: I8946161064411f5b7cedf33b6669d0886789222e