Run PackageManagerShellCommandInstallTest in presubmit
ag/25462013 seperated tests to several small tests but it forgot
to install test apk in testGetProperty. The test should fail and
can be captured in presubmit. But we moved to postsubmit to avoid
SLO issue.

Ideally we should run CtsPackageManagerTestCases in presubmi but it
is possible to introduce SLO issues back. We try to add min set
PackageManagerShellCommandInstallTest first.

Bug: 314222113
Test: atest --test-mapping cts/tests/tests/content
Change-Id: I366f133f2c552d2620618808dd81fd253969cc90