Consistently include all `util` tests.
We've been working to progressively opt-in tests for util classes,
and we've now crested the hill, so let's include all tests by default
and opt-out the handful of remaining classes that aren't supported
under Ravenwood.

This helps us future-proof to make it easier to ensure that all
newly added tests are being run on both devices and Ravenwood.

Bug: 292141694
Test: atest CtsUtilTestCasesRavenwood CtsUtilTestCases
Change-Id: I86a5dd2e5c2bb78f1713721c2f1a440dcd7f358b