mediacuj CTS: Fix Orientation test failure
In setActivity() function we are calling setRequestedOrientation(x) to
set the orientation and then calls getDeviceOrientation() to initialize
current orientation but on slower device like cf
setRequestedOrientation() is taking more time to take the effect which
leads to setDeviceOrientation() doesn't returns value passed (i.e. x)
in setRequestedOrientation() due to which tests fails.
Now the test uses 'x' value to initialize current orientation and then
calls setRequestedOrientation(x) to set orientation. Later after
orientation change is done we are calling getDeviceOrientation() to
get the current orientation and verify that it should not match with
previous orientation.

Also, modified tests to verify the orientation change after each
orientation change call, earlier orientation change was getting
verified only once for a playlist.

Bug: 310783217

Test: atest CtsMediaCUJSmallTest

Change-Id: I7cf4972d7fb247b0b6b820f90d6bd381e2bec511