Fix test crash due to SurfaceControl not being ready in onAttach
There's a bug where the onAttachedToWindow doesn't necessarily mean the
SurfaceControl has been created yet. b/286406553 is opened tracking that
issue but for now fix the test so it doesn't crash. If the
SurfaceControl is not yet ready in attach, we wait for the first draw to
run the test.

Fixes: 315634855
Bug: 286406553
Test: AttachedSurfaceControlSyncTest
Change-Id: Ifd64c356a190fe41c8d8d00dbb9d93c9358ce71e