CTS for race condition and minor fix
Replace singleThreadExecutor that is being used for
VisualQueryDetectionService's tests to simulate more general executor
(since the same executor is used for file reading too). Also added a
test case to simulate the case where multiple streamQuery() and
finishQuery() ot be called, the signal from the VIS should be processed
sequentially.

Bug: 295390470
Test: atest CtsVoiceInteractionTestCases
Flag: N/A
Change-Id: Iad5d5628a3e3c9e8d73af5c320d6144cdb73ab96