Fix HotwordDetectionService training data tests for HSUM configuration.
Currently the hotword detection service training data tests involve
enabling the training data app-op for the shell identity to allow
training data egress from the trusted process. Currently these tests
grant the app-op to the shell uid tied to the system user, assuming that
these tests are running for the system user,
This assumption does not work in HSUM (go/supporting-hsum)
configuration, which Android Auto and Tangor Tablets operate in.

This CL fixes this assumption by micming the shell uid calculation logic in ShellDelegate within the ActivityManagerService to make sure the app-ops are granted to the shell uid for the proper user.

Test: atest -c
CtsVoiceInteractionTestCases:HotwordDetectionServiceBasicTest
Bug: 310313240
Change-Id: I98611ea099d69d8dbea476f689474dca2a6d7513