Baseline NewApi issues
NewApi is a lint check that you don't call framework methods that were
introduced in versions later than your min_sdk_version. We want to
make this an error, so we're baselineing all existing issues.

Bug: 268261262
Test: Presubmits
Ignore-AOSP-First: It's too difficult to do this LSC across multiple branches, so we're only doing it in internal main
Change-Id: I451206d3d0aa816d65f71bd89d9f4fa3c1fdbe67