add CTS coverage for TM#getRegisteredPhoneAccounts
currently there is no way for VoIP applications to fetch their
registered PhoneAccounts with Telecom. TelecomManager#getPhoneAccount is
gated by the READ_PHONE_STATE permission which is a higly privileged
permission that VoIP apps will not be granted. This API will allow VoIP
apps to get a registered PhoneAccount with Telecom which is helpful to
know what calling capabilites are currently registered etc.

Bug: 307609763
Test: CTS coverage
Change-Id: Iadd5855109e4f2b6b82f9cc8aabd1755c1177a3c