Move router instantiation from setup to test methods
This change allows individual tests to control when MediaRouter2 is
instantiated, so that they can call adoptShellPermissionIdentity()
before MediaRouter2 instantiation. This is important for permissions
that are only checked during instantiation, because they are not
expected to change during the lifecycle of the router owner, like
MEDIA_ROUTING_CONTROL.

Modifying permission access directly from the test methods works because
these are invoked as host-side tests, which means that each individual
test method will run isolated in its own process. This isolation prevents
accidental state sharing in MediaRouter2.

Bug: 205124386
Test: atest MediaRouter2HostSideTest
Change-Id: I5e28b3f2dbcda9c08ad5f2cb65b31af012d613b1