fix test failure when DUT does not have another CE to switch to
A test failure came in for testBasicAudioSwitchTest.  This test was only
failing on Tangor devices which makes sense because the SPEAKER audio
rotue is the only available audio route on those devices. Normally,
pixels have the EARPIECE route that can be used to switch to for
testing.  This test should be bypassed when the DUT does not have a
non-null endpoint to switch to.

Bug: 321075762
Test: CTS test improvement
Change-Id: If1a7bfd0957d37dec7f1ccf2fb6a6ac4a2415e17