Simplify MultiUserTest with Bedstead
This is a follow up CL to our previous CL [1], which migrated
MultiUserTest from CtsInputMethodServiceHostTestCases to
CtsInputMethodInstallTestCases.

This CL removes a lot of boilerplate code by relying on Bedstead. As
they promised, the code is now much simpler and more descriptive with
reducing the total test execution time from 57s to 35s.

 [1]: I9306fd65d90e203238ae349f35b7cc2b46b4562e
      794b78a2561e3b63e5fd742aa04cade10604c96a

Bug: 283765791
Fix: 323240011
Test: atest CtsInputMethodInstallTestCases
Test: atest CtsInputMethodInstallTestCases --instant
Change-Id: I14dc1368d1a26b17c81a616a62431cf0b2c85de4