Update GameFrameRate test for all refresh rate divisors
Previously we only test the divisors under the same display mode
refresh rate, which does not really cover all the behaviors. To improve
this, this patch updates the test cases to override all divisors under
all display modes, and when the frame rate we are trying to override is
not a certain refresh rate, we will choose the closest possible frame
rate under that refresh rate as the expected one.

Furthermore, this patch also updated the test case for default game
frame rate. It checks the sysprop as well as observing the default frame
rate is in effect.

Bug: 323105276
Bug: 321179619

Test: atest CtsGameFrameRateTestCases
Change-Id: I0bfd5165cd8ecca86ed205b240dff4746bdfb057