InputAtomsTest: check repeatedly for reported atom
Most of the time, sleeping for AtomTestUtils.WAIT_TIME_LONG (1 second)
and checking once for a reported TouchpadUsage atom is sufficient in
this test, but sometimes (I'd guess around 2–3% of the time), the atom
hasn't been reported by that time, causing flakes. In my tests, sleeping
for 5 * WAIT_TIME_LONG always picked up the atom at the cost of 4
seconds increased runtime. Instead, keep sleeping for WAIT_TIME_SHORT
(0.5 seconds) and check after every sleep, up to a maximum of 10 times.
This should fix this kind of flake while reducing the run time of
average runs.

Bug: 324047709
Test: on cf_x86_64_phone-trunk_staging-userdebug target, run:
    $ atest 'android.input.cts.hostside.InputAtomsTest#testTouchpadUsageAtom_FingerAndPalmCounts' \
            'android.input.cts.hostside.InputAtomsTest#testTouchpadUsageAtom_TwoFingerSwipeCount' \
	    'android.input.cts.hostside.InputAtomsTest#testTouchpadUsageAtom_ThreeFingerSwipeCount' \
	    'android.input.cts.hostside.InputAtomsTest#testTouchpadUsageAtom_FourFingerSwipeCount' \
	    'android.input.cts.hostside.InputAtomsTest#testTouchpadUsageAtom_PinchCount' \
	    --iterations=100
Change-Id: I392626dc33fa0ad65d161578b09f7da13c40f907