Make btest resilient to ADB disconnection.
This moves to using nohup on the device and using files to communicate
the test logs. This means adb can be disconnected during the test - the
test will continue to run - and then we can reconnect to collect the
results.

This will be used in a follow up CL to enable untethered testing as an
alternative to using adb-over-wifi.

Ideally we will add similar support to Tradefed to offer this
alternative to partners - but for the time being it can be used for
development at least.

Change-Id: Iee99b327a3c0b8ae5d64b315b13d4c37b52d8ba3