Fix failing AudioFocusTest on Automotive builds
Tests that are annotated with @RequireNotAutomotive are
expected to be skipped on automotive builds. Due to the
missing device state, these tests are falsely executed
on automotive builds and result in failure.

Fix the test by creating device state. Now these tests
are skipped if the device state matches FEATURE_AUTOMOTIVE.

Bug: 322385482
Test: atest AudioFocusTest
Change-Id: Ic9360e01d9dfc323ad77f591c1f8fdd8d8ac4a1f