Camera ITS: Initial cl for multi camera switch.
This change includes adding a new test to verify the behavior
of the camera when the lens changes from UW to W.
It checks for the skip conditions, find out the crossover
point from UW to W during dynamic preview recording with
different zoom ratios.
This test is added under a new scene directory scene7.

Bug: 308770424
Test: Tested locally on pixel device.

Change-Id: I2a88823f209b26307cf57064d5355c079744d2e3