[Autofill Framework] Add STS test to verify autofill service checks intent-filter when app updates
The new test testAutofillServiceDisabledAfterUpdateRemoveIntentFilter()
verify it by first enabling the TestAutofillService (which has the
intent filter) and verify current autofill service is correctly set.
Then it installs an update version of TestAutofillService apk but with the
intent-filter for autofill service removed in its AndroidManifest. After
the update, the test verifies autofill service disabled.

To compare with, the current existing test testHasEnabledAutofillServices()
is checking autofill is still enabled after a normal autofill service
updates (with correct premission and intent-filter set in the update apk).

Bug: b/324874908
Test: atest CtsAutofillServiceTestCases
Change-Id: Ib6960f977e20ec61e9d85f8563e1d96d734a072d