Fix GameFrameRateTest failures
Adding flag annotations to the test cases to make
sure that they test with the required flags enabled.

In addition, this test also add testing logic for
game default frame rate. Previously this test only verifies
game frame rate overrides from game interventions.
Adding the section of removing game interventions and checking
the default frame rate is set to make sure its sysprop is set
as well as the behavior is as expected.

Bug: 327018158
Bug: 323417524
BUg: 323105276
Test: atest CtsGameFrameRateTestCases
Change-Id: Ibaa0490727afd4fdf9bcf5ac6d90e1b254580e65