Don't reply on the APPLY_RESTRICTION flag in tests.
The APPLY_RESTRICTION flag was merely the same thing as not having any
EXEMPT flags for a restricted permission. We are re-purposing that flag
to reflect the actual restricted state of the permission, especially the
soft restricted permissions, in order to simplify various permission
logic, so just make the test stop checking the APPLY_RESTRICTION flag
while it still checks the EXEMPT flags which is exactly the same thing.

Bug: 266164193
Test: PackageSetInstallerTest
Change-Id: I0328581c8d51ab46c18de7230bb4849e3060be4d